Split internal and public API tests
